+++ src/Twig/Extension/RecordExtension.php
@@ -23,7 +23,6 @@ class RecordExtension extends Extension
     {
         $safe = ['is_safe' => ['html']];
         $env  = ['needs_environment' => true];
-        $deprecated = ['deprecated' => true];
 
         return [
             // @codingStandardsIgnoreStart
@@ -32,7 +31,6 @@ class RecordExtension extends Extension
             new TwigFunction('fields',        [Runtime\RecordRuntime::class, 'fields'], $env + $safe),
             new TwigFunction('listtemplates', [Runtime\RecordRuntime::class, 'listTemplates']),
             new TwigFunction('pager',         [Runtime\RecordRuntime::class, 'pager'], $env + $safe),
-            new TwigFunction('trimtext',      [Runtime\RecordRuntime::class, 'excerpt'], $safe + $deprecated + ['alternative' => 'excerpt']),
             // @codingStandardsIgnoreEnd
         ];
     }
@@ -43,14 +41,12 @@ class RecordExtension extends Extension
     public function getFilters()
     {
         $safe = ['is_safe' => ['html']];
-        $deprecated = ['deprecated' => true];
 
         return [
             // @codingStandardsIgnoreStart
             new TwigFilter('current',     [Runtime\RecordRuntime::class, 'current']),
             new TwigFilter('excerpt',     [Runtime\RecordRuntime::class, 'excerpt'], $safe),
             new TwigFilter('selectfield', [Runtime\RecordRuntime::class, 'selectField']),
-            new TwigFilter('trimtext',    [Runtime\RecordRuntime::class, 'excerpt'], $safe + $deprecated + ['alternative' => 'excerpt']),
             // @codingStandardsIgnoreEnd
         ];
     }